+++ client/src/shared/apis/errorhandlers/index.ts
@@ -1,3 +1,4 @@
 export { errorManager } from './errorManager';
 export { serverDisconnected } from './serverDisconnected';
 export { reRequest } from './reRequest';
+export { tokenReissue } from './tokenReissue';
