+++ client/src/shared/hooks/index.ts
@@ -4,9 +4,8 @@ export { default as useMediaRecord } from './record/useMediaRecord';
 export { default as useModal } from './modal/useModal';
 export { default as usePwCheck } from './pwcheck/usePwCheck';
 export { default as useSectionObserver } from './sectionobserver/useSectionObserver';
-export { default as useThemeToggler } from './useThemeToggler';
+export { default as useThemeToggler } from './theme/useThemeToggler';
 export { default as useToast } from './toast/useToast';
 export { default as useWebSpeech } from './speech/useWebSpeech';
-export { default as useDragCorrection } from './useDragCorrection';
 export { default as useInterval } from './interval/useInterval';
 export { default as useCounter } from './counter/useCounter';
